pyproject.toml
[tool.poetry]
name = "appcfg"
version = "0.1.0"
description = "Flexible hierarchic Python application configuration"
authors = ["bjoluc <mail@bjoluc.de>"]
license = "MIT"
readme = "readme.md"
homepage = "https://github.com/bjoluc/appcfg"
repository = "https://github.com/bjoluc/appcfg"
documentation = "https://github.com/bjoluc/appcfg#readme"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: Implementation :: CPython",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Topic :: Software Development :: Libraries",
    "Intended Audience :: Developers",
    "Typing :: Typed"
]
keywords = [
    "app",
    "application",
    "config",
    "configuration",
    "env",
    "environment",
    "node-config",
]

[tool.poetry.dependencies]
python = "^3.6"
deepmerge = "^0.1.0"
pyyaml = { version = "^5.3.1", optional = true }

[tool.poetry.extras]
yaml = ["pyyaml"]

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-watch = "^4.2.0"
pytest-voluptuous = "^1.1.0"
pyfakefs = "^4.0.2"
flake8 = "^3.8.2"
black = "^19.10b0"
doc2md = "^0.1.0"
coverage = { extras = ["toml"], version = "^5.1" }
commitizen = "^1.22.0"

[tool.coverage.run]
branch = true
include = ["appcfg/*", "tests/*"]

[tool.commitizen]
version = "0.1.0"
version_files = ["pytest_reraise/__init__.py", "pyproject.toml:version"]
bump_message = "chore(Release): Release $new_version [skip ci]"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
